significant refactoring:
 - add search-helper class
 - add result pagination
 - add search result meta info
 - add try/catch for better ajax error handling
 - add confirm on populate/clear buttons
 - make populate and clear ajax calls use post method

+++ ajax/clear-comments.php
@@ -7,15 +7,16 @@ require '../include/mysql-connect.php';
  * all we're doing here is deleting all records in our comments table
  */
 
-$response = null;
-$result = $con->query('delete from comments');
+if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+	$result = $con->query('delete from comments');
 
-if ($result) {
-	$response = new AjaxResponse(0, 0, $con->affected_rows,
-		ResponseStatus::Ok);
-} else {
-	$response = new AjaxResponse(0, 0, 0, ResponseStatus::Error,
-		$con->error);
+	if ($result) {
+		$response = new AjaxResponse(0, 0, $con->affected_rows,
+			ResponseStatus::Ok);
+	} else {
+		$response = new AjaxResponse(0, 0, 0, ResponseStatus::Error, null,
+			$con->error);
+	}
+
+	$response->send();
 }
-
-$response->send();
